+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- PHP 7.1+ is now required
+- Optimized algorithm for getting unique random integers
 
 ## [3.0.0] - 2017-04-16
 ### Changed